+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add Display Header option (#15)
 - Add the ability to collapse parent rows (#16)
 - Add a Sticky option to follow when scrolling (#17)
+- Add Table Filter (#18)
 
 ## 1.2.0 (2023-06-26)